added datalayer necessary for sending events to GA
